Refactor: stateSetter 사용 추상화
setStateOf으로 감싸고 에러 처리 추가
