+++ src/utils/Errors/SetStateError.js
@@ -1,0 +1,6 @@
+export default class SetStateError extends Error {
+  constructor(message) {
+    super(message);
+    this.name = "SetStateError";
+  }
+}
