+++ src/utils/constants.js
@@ -20,6 +20,7 @@ export const ERROR = {
   INVALID_HOME_STATE: "Tried to set home's state using invalid data.",
   NONE_ARRAY_STATE: "Tried to set state using none array value.",
   INVALID_REQUEST: "서버에 문제가 발생했습니다.\n 잠시 후 다시 시도해주세요.",
+  UNREGISTERED_SETTER: "Tried to use unregistered state setter.",
 };
 
 export const EVENT = {
